# contracts/openapi-v1.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ConverterRequest" : {
        "description" : "Request to convert images to calendar events",
        "type" : "object",
        "required" : [ "files" ],
        "properties" : {
          "files" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImageFile"
            },
            "description" : "List of image files to convert",
            "maxItems" : 10,
            "minItems" : 1
          },
          "timeZone" : {
            "type" : "string",
            "description" : "Timezone for event times (e.g., 'America/New_York')",
            "examples" : [ "UTC" ],
            "pattern" : "[A-Za-z0-9/_+\\-]+"
          },
          "currentDate" : {
            "type" : "string",
            "description" : "Current date in ISO-8601 format for context",
            "examples" : [ "2024-01-31" ],
            "pattern" : "\\d{4}-\\d{2}-\\d{2}"
          },
          "userId" : {
            "type" : "string",
            "description" : "User identifier for quota tracking"
          }
        }
      },
      "ConverterResponse" : {
        "required" : [ "success", "icsContent" ],
        "description" : "Successful image-to-calendar conversion",
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean",
            "description" : "Whether the conversion was successful",
            "examples" : [ true ]
          },
          "icsContent" : {
            "type" : "string",
            "description" : "The generated ICS calendar content",
            "examples" : [ "BEGIN:VCALENDAR..." ]
          },
          "error" : {
            "type" : "string",
            "description" : "Error message if conversion failed"
          },
          "message" : {
            "type" : "string",
            "description" : "Detailed error description"
          },
          "details" : {
            "description" : "Additional error details"
          },
          "warnings" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "description" : "Non-blocking review warnings produced during calendar normalization"
          }
        }
      },
...
      "ErrorResponse" : {
        "required" : [ "success", "error", "message", "errorCode", "timestamp", "status", "requestId" ],
        "description" : "Canonical API error",
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "details" : { },
          "timestamp" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
      "ImageFile" : {
        "type" : "object",
        "description" : "Image file to be processed",
        "properties" : {
          "dataUrl" : {
            "type" : "string",
            "description" : "Base64 encoded image data URL",
            "examples" : [ "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ..." ]
          },
          "url" : {
            "type" : "string",
            "description" : "URL to a publicly accessible image"
          }
        }
      },
...
      "PlanType" : {
        "type" : "string",
        "enum" : [ "FREE", "PRO", "BUSINESS", "UNLIMITED" ]
      },
      "Quota" : {
        "type" : "object",
        "required" : [ "usageCount", "limit", "remaining", "plan" ],
        "description" : "Quota usage for the active monthly period",
        "properties" : {
          "usageCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remaining" : {
            "type" : "integer",
            "format" : "int64"
          },
          "plan" : {
            "$ref" : "#/components/schemas/PlanType"
          }
        }
      },
      "QuotaStatusResponse" : {
        "required" : [ "success", "quota", "enabled" ],
        "description" : "Public quota status",
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "quota" : {
            "$ref" : "#/components/schemas/Quota"
          },
          "enabled" : {
            "type" : "boolean"
          }
        }
      },
      "Statistics" : {
        "description" : "Global conversion statistics aggregated from Notion tracking database",
        "type" : "object",
        "properties" : {
          "fileCount" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Total number of image files processed in successful conversions"
          },
          "eventCount" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Total number of calendar events generated in successful conversions"
          }
        }
      },
...
    }
  },
...
  "paths" : {
    "/v1/converter" : {
      "post" : {
        "summary" : "Convert images to calendar events",
        "description" : "Uses AI to extract calendar events from images and convert them to ICS format",
        "tags" : [ "converter" ],
        "parameters" : [ {
          "name" : "X-Installation-ID",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConverterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Conversion successful",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConverterResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request data",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Verified Google authentication required",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Idempotency conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Processing error - valid input but conversion failed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Quota or rate limit exceeded",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502" : {
            "description" : "AI provider rejected the request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "Quota datastore unavailable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504" : {
            "description" : "AI provider timed out",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/converter/quota-status" : {
      "get" : {
        "summary" : "Get user quota status",
        "description" : "Retrieves the current quota usage and plan information for a user",
        "tags" : [ "converter" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          },
          "required" : true
        }, {
          "name" : "X-Installation-ID",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Quota status retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuotaStatusResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Verified Google authentication required",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/v1/converter/statistics" : {
      "get" : {
        "summary" : "Get usage statistics",
        "description" : "Retrieves usage statistics and analytics data",
        "tags" : [ "converter" ],
        "responses" : {
          "200" : {
            "description" : "Statistics retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Statistics"
                }
              }
            }
          },
          "502" : {
            "description" : "Failed to fetch statistics from external service"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
...
  },
  "info" : {
    "title" : "3dime-api with Quarkus",
    "version" : "1.0",
    "description" : "This is a Microservice 3dime-api based on the OpenAPI 3.0 specification. Click [here](https://github.com/m-idriss/3dime-api). You can load via the `Code > Clone`, enjoy.",
    "contact" : {
      "name" : "Idriss",
      "email" : "contact@3dime.com",
      "url" : "https://github.com/m-idriss"
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}